+++ src/scruby/mixins/update.py
@@ -9,19 +9,14 @@ from __future__ import annotations
 __all__ = ("Update",)
 
 import concurrent.futures
-import logging
 from collections.abc import Callable
-from typing import Any, TypeVar
+from typing import Any
 
 import orjson
 from anyio import Path
 
-logger = logging.getLogger(__name__)
 
-T = TypeVar("T")
-
-
-class Update[T]:
+class Update:
     """Methods for updating documents."""
 
     @staticmethod
@@ -30,7 +25,7 @@ class Update[T]:
         filter_fn: Callable,
         hash_reduce_left: str,
         db_root: str,
-        class_model: T,
+        class_model: Any,
         new_data: dict[str, Any],
     ) -> int:
         """Task for find documents.
@@ -91,7 +86,7 @@ class Update[T]:
         update_task_fn: Callable = self._task_update
         hash_reduce_left: int = self._hash_reduce_left
         db_root: str = self._db_root
-        class_model: T = self._class_model
+        class_model: Any = self._class_model
         counter: int = 0
         with concurrent.futures.ThreadPoolExecutor(max_workers) as executor:
             for branch_number in branch_numbers: